Fix 100% CPU usage in asyncore.loop().
The default implementation of asyncore.dispatcher.writable() returns
True, which means that dispatcher's socket gets added to the writable
list that is pased to select(). This in turn means that asyncore.loop()
will consume all available CPU when there is no activity on any socket,
by repeatedly calling select() with no delay, as it is constantly
notified that the dispatcher's socket is writable.

Also, remove the timeout parameter to asyncore.loop(), since it's
not needed.