Stop trying to parse packets after error.
If we encounter an unexpected error during parsing,
then we probably exposed a bug in a message definition
somewhere and we're out of sync in the message stream.
Most likely everything afterwards will be wrong, so reset
the stream to before the packet that triggered the error,
and then just pass the data through blindly.