Add --log-file option.
Since we plan to distribute this as an egg, we can't store
a log configuration in the same dir as mc3p.util because we
might not have permissions.

With this commit, we use a basicConfig at ERROR level by default.
If the user specifies the --log-file option, we use the specified
logfile (and create it if it doesn't exist).