Platform-independent plugin message injection.
The original mechanism for plugin message injection
used UNIX sockets and asyncore to asynchronously send messages from
a plugin to an mc3p proxy. This obviously doesn't work in Windows.

This commit replaces the UNIX socket connections with
multiprocessing.Queue objects, which use a pipe and backing
thread to implement a process-safe queue.