Replace 'dst' and 'dir' args with 'source'.
Some of the PluginManager code incorrectly used the name
'dst' for the argument indicating message direction. That
argument indicates the *sender*, not the *receiver*. Similarly,
the 'dir' argument name is ambiguous. Both have been replaced with
'source'.

In the process, this fixes a minor bug that was caused by the
above mixup: plugins should be instantiated after the *server*
sends msgtype 0x01, not the client. The tests have been updated
to reflect this.